+++ src/Commands/ThreadHandleCommand.cs
@@ -1,5 +1,4 @@
 using System;
-using System.ComponentModel;
 using System.Management.Automation;
 using System.Runtime.InteropServices;