+++ tests/TokenContextCommand.Tests.ps1
@@ -1,0 +1,12 @@
+$moduleName   = (Get-Item ([IO.Path]::Combine($PSScriptRoot, '..', 'module', '*.psd1'))).BaseName
+$manifestPath = [IO.Path]::Combine($PSScriptRoot, '..', 'build', $moduleName)
+
+Import-Module $manifestPath
+
+Describe "Enter-TokenContext" {
+
+}
+
+Describe "Exit-TokenContext" {
+
+}